+++ packages/react-compass/src/index.ts
@@ -1,3 +1,4 @@
+export {SSRProvider} from '@react-aria/ssr'
 export {default as Accordion} from './accordion'
 export type {
   AccordionProps,
@@ -171,6 +172,12 @@ export {default as Radio} from './radio'
 export type {RadioGroupProps, RadioProps} from './radio'
 export {default as RangeCalendar} from './range-calendar'
 export type {RangeCalendarProps, RangeValue} from './range-calendar'
+export {default as RichTextEditor} from './rich-text-editor'
+export type {
+  Content,
+  JSONContent,
+  RichTextEditorProps,
+} from './rich-text-editor'
 export {default as RichTextRender} from './richtext-render'
 export type {Document, RichTextRenderProps} from './richtext-render'
 export {default as SearchField} from './searchfield'